Create GUID class, add network rejoin and peer fallback mechanisms
GUID is a wrapper class instead of using integers
to represent globally unique IDs. It allows us to
attach GUID specific methods to the object.

Allow clients to rejoin the network using previously
allocated GUID. When a client initially adds the
GUID to redis it is also written to disk. When a
client initially attempts to load it's GUID from
redis, if one is not set, the client will check for
a file containing the previously allocated GUID.

Add fallback mechanism where if a primary peer is not
responsive, a client can attempt to connect to another
peer on the network.