Give Node class access to database
It's messy to pass the db instance around to Node
methods, but it's becoming clear that they will need
access to the database. A downside I do not like is
that you have to remember to set the class attribute.
Also made boot_node a Node class attribute to remove
the `settings` import even though it has the same
drawback of it needed to be explicitly set.